Rename file and refactor its location
